Percobaan cart

+++ lib/models/product.dart
@@ -4,7 +4,7 @@ class Product {
   final String? id;
   final String? title;
   final String? description;
-  final double? price;
+  final int? price;
   // final int? qty;
   final String? image;
   bool isFavorite;
